backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.21
python-dotenv==1.2.1
pandas==2.3.3
email-validator==2.3.0
PyPDF2==3.0.1
dnspython==2.4.2
